Change paths so that they are more inline with proper API conventions. Move Interceptor config to config package
