+++ src/main/java/application/config/InterceptorConfig.java
@@ -1,4 +1,4 @@
-package application;
+package application.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
@@ -11,7 +11,7 @@ public class InterceptorConfig implements WebMvcConfigurer {
 
     @Override
     public void addInterceptors(InterceptorRegistry registry){
-        registry.addInterceptor(new HostAuthInterceptor()).addPathPatterns("/host");
+        registry.addInterceptor(new HostAuthInterceptor()).addPathPatterns("/host/rooms");
         registry.addInterceptor(new RoomExistenceInterceptor())
         .addPathPatterns("/rooms")
         .addPathPatterns("/rooms/*")
